Decrease size of cursor handle and increase size of its touch target
Summary:
Previously, the cursor handle was too small to reasonably interact with on mobile devices. We then
increased its size to ~40px x 40px. However, this made it look oddly large compared to the math
input. In this diff, I've shrunk down the cursor to ~22px x 22px (per Nefaur's recommendation) and
increased its touch target to 44px x 44px.

Additionally, your finger now locks to the bottom of the cursor, so that you can see it as you
drag. Previously, the cursor would be invisible under your finger. This can make for a slightly
awkward effect if the user starts the touch near the top of the cursor, as it will then jump up
above the math, but it seems like a better overall user experience than before.

On Asana @ https://app.asana.com/0/133656453415343/133227990069246.

Test Plan:
- Open up `index.html`.
- Type some math. Tap into the input to reveal the cursor handle.
- Open up Devtools.
    - Verify that the cursor is ~22px x 22px (it's a little taller than it is wide, as it's not a perfect square).
    - Verify that its bounding box (the `<span>`) is 44px x 44px.
    - Verify that the cursor is perfectly centered (horizontally) in the bounding box.
- Interact with the cursor by dragging it.
    - Verify that the bottom of the cursor locks to your finger on move.

Reviewers: benkomalo

Reviewed By: benkomalo

Subscribers: nefaurk, kevinb

Differential Revision: https://phabricator.khanacademy.org/D27393