Scroll input into view when obscured on focus
Summary:
Previously, whenever an input was focused, the page would auto-scroll such that it appeared 16px
from the top. This made for a frustrating user experience whereby parts of the question might be
obscured after focusing an input. This diff lays the foundations for better auto-scroll behavior;
these foundations will be put to use in an upcoming Perseus diff.

Specifically, the new auto-scroll behavior is as follows:

- On focus:
    - If the keypad would obscure the input, auto-scroll such that the input appears 16px above the keypad.
    - If the input is outside of the viewport, auto-scroll such that the input appears 16px from the top of the viewport. (This can't happen on mobile web, as you can only focus an input by tapping it; but on native, you might navigate to an input with the external arrows.)

On Asana @ https://app.asana.com/0/133656453415343/134290939887000.

Test Plan: Coming in the Perseus integration diff.

Reviewers: benkomalo

Reviewed By: benkomalo

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27408