Unbind mysterious MathQuill mousedown handler
Summary:
This one, I can't really explain. MathQuill binds a `mousedown` handler to its container element.
This handler makes the cursor appears and also allow the user to drag to select more text. However,
this handler can sometimes be triggered by tapping slightly outside of the bound container--maybe
5px or so in any direction. As a result, the cursor appears, but we don't get any `mouseDown` or
`touchStart` event on our own container and, as such, don't trigger a focus event, leaving the
cursor active but the keypad dismissed.

The fix, it appears, is to unbind the handler. Note that this handler wasn't being triggered for
normal taps on the input--only for taps near the edges. Further, we don't need any of those
behavior as we manage all the cursor interactions ourselves.

On Asana @ https://app.asana.com/0/133656453415343/134437167943898.

Test Plan:
- Open up the demo page.
- Enable mobile emulation.
- Tap on the input.
    - Verify that the keypad appears, and the cursor appears.
- Tap off the input.
    - Verify that the keypad disappears, and that the cursor disappears.
- Tap within 5px of the edge of the input.
    - Verify that nothing happens--no cursor appears, etc.
- Focus the input, type a fraction, hit delete (to enable selection).
- Tap elsewhere.
    - Verify that the cursor moves, selection clears, handle appears, etc.
    - Verify that the cursor is draggable as before.

Reviewers: benkomalo

Reviewed By: benkomalo

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27429