Replace SVG numeral icons with unicode symbols
Summary:
This will cut down the size of the payload a bit. I was hoping that I could be a bit more
aggressive by replacing some of the other symbols (like the > and < symbols), but some of those are
customized and not Proxima-defaults.

On Asana @ https://app.asana.com/0/133656453415343/135753965927511.

Test Plan:
- Open up the keypad.
    - Verify that the numerals look okay.
    - Verify that the numerals are not italicized, but the symbols (variables + Pi) are.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D27489