Fill in unfortunate cracks between keypad buttons
Summary:
This is an unfortunate diff. On some devices, sometimes, there are gaps between the borderless
keypad buttons, apparently due to rounding errors or subpixel rendering issues. Here, I've added
the proper background colors to the various keypads to try and hide the worst of the issues.
Unfortunately, there are still some cases where seepage could be possible (e.g., in between the
dismiss and backspace buttons in the simpler keypads), but I can't find a better way to handle
this.

On Asana @ https://app.asana.com/0/133656453415343/135512699771133.

Test Plan:
- Open up the Fraction and Number keypads.
    - Verify that there are no visible cracks between the numeral keys.
- Open up an Expression keypad.
    - Verify that there are no visible cracks between the numeral keys and between the keys in the sidebar.
    - Drag the keypad a bit.
        - Verify that the area behind the keypad is still dark grey.

Reviewers: benkomalo

Reviewed By: benkomalo

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27468