Disable `appear` animation for keypad
Summary:
This is unnecessary and was causing some additional jank in the keypad animation.

There is still a ton of jank in the animation, though. It's somewhat inconsistent. Most of it
seems to come from jQuery layout thrashing, some of which is being instigated by MathQuill.

Test Plan: Verify that the keypad still animates in and out as before.

Reviewers: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D27467