Make fraction behavior configurable
Summary:
Currently, if you hit the fraction key, it will take the math to the left of your cursor and put
that in the numerator of a new fraction, with an empty denominator. We want to test a variant of
this behavior that always creates an empty fraction at the cursor location, and leaves the existing
math intact. This diff makes that behavior customizable via the `/custom.html` creator.

Test Plan:
- Open up the keypad.
- Type some numbers.
- Type a fraction.
    - Verify that the numbers go in the numerator.
- Go to `custom.html`.
- Choose "inclusive" for the "fraction behavior".
    - Verify that the resulting behavior is as before.
- Go to `custom.html`.
- Choose "exclusive" for the "fraction behavior".
- Type some numbers.
- Type a fraction.
    - Verify that the fraction appears at the cursor location with an empty numerator and denominator.

Reviewers: benkomalo

Reviewed By: benkomalo

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27499