Prevent keypad components from accessing store directly
Summary:
This is the first in a series of diffs aimed at refactoring `math-input` to _not_ rely on a
singleton, globally-accessible store.

In this diff, I've removed all the direct accesses to the store that were being made by child
components of the keypad, primarily by relying instead on the `mapDispatchToProps` method proivded
by `react-redux`.

Test Plan:
- NFC--just a refactor.
- Open up the keypad demo.
- Open up the keypad.
    - Verify that you can still swipe between pages.
    - Verify that everything is sized properly.

Reviewers: kevinb, jared

Reviewed By: jared

Subscribers: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D27840