Disable left and right arrows based on cursor position
Summary:
With this diff, we now disable the left and right navigation arrows when the cursor is located at
the left and right ends of the input.

In doing so, it introduces three new cursor contexts: `LEFT_END`, `RIGHT_END`, and `EMPTY`, which
represent being positioned at the left- and right-ends of the input, or both (i.e., in an empty
input). Note that these three contexts all imply `TOP_LEVEL`.

For now, when the buttons are "disabled", they don't render a focus state, though we may want to
modify the iconography as well.

On Asana @ https://app.asana.com/0/114493820703878/140146213657105.

Test Plan:
- Open up the keypad in the demo page.
    - Verify that tapping the arrows does not trigger any highlight feedback.
- Type some math; keep your cursor at the right end.
    - Verify that the right arrow does not respond to touches.
    - Verify that the left arrow does respond to touches.
- Drag your cursor to the left end.
    - Verify that the left arrow does not respond to touches.
    - Verify that the right arrow does respond to touches.

Reviewers: jared, benkomalo

Reviewed By: benkomalo

Subscribers: shadaj

Differential Revision: https://phabricator.khanacademy.org/D27874