Use `window.innerHeight` when computing scroll offset
Summary:
Previously, to compute whether or not we needed to scroll a focused input into view, we used
`document.body.clientHeight`. However, in the mobile apps, this could actually return a value
larger than the height of the viewport, since often the page is made scrollable by a native
container, with the `body` allowed to extend beyond the height of the window. Instead, we need to
use `window.innerHeight`, which correctly returns the height of the viewport.

Test Plan: Same as D27408. Native testing will come in the integration diff.

Reviewers: benkomalo

Reviewed By: benkomalo

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D28045