Revert to setting keypad position via `bottom: 0`
Summary:
In D28058, we changed the keypad positioning logic to set a `top` value based on
`window.innerHeight`. However, in testing, its been revealed that `window.innerHeight` does not
account for the presence of additional browser chrome and can thus change as that chrome appears
and disappears. For example, on my iPhone 5s in Safari, `window.innerHeight` is reported as 460px
with the bottom bar visible and 529px without. As a result, the keypad would be anchored with its
top pinned to the top of the bottom bar; when scrolling, the keypad would become visible, even
though it was inactive.

Test Plan:
- Substate into webapp.
- Scroll around in Safari.
    - Verify that the keypad is never visible.
    - Verify that it appears on tap, and that it focuses properly.

Auditors: benkomalo, kevinb, shadaj