Move dismissal and activation animation into container component
Summary:
This diff moves the dismissal container a level above the actual keypad DOM nodes, such that the
same container persists across re-configurations. This means that when a keypad of type X is
dismissed and inactive, and a user taps on an input that re-configures the keypad to be of type Y,
the same dismissal container will be used, with the contents replaced before the animation is
triggered.

Re-using the same container is important, as we rely on a CSS transition. If we replace the
dismissal container with an active container in-full, we won't be transitioning from dismissed to
active; we'll just be rendering the active keypad from the start.

In practice, we'll now see the keypad animation into place appropriately on first focus.

On Asana @ https://app.asana.com/0/132409062806753/143244203382661.

Test Plan:
- Verify that all dismissal behavior is the same as before in the local demo.
- Substate into webapp.
    - Verify that, on first focus for inputs with different types of keypads, the keypads animate in correctly.
    - In particular, navigate to an exercise that uses non-Expression inputs (e.g., `/e/compound-events`).
        - Verify that, on focus, the keypad animates into place appropriately.
- (May need to test on an actual device; I wasn't seeing the bad behavior in the Chrome emulator.)

Reviewers: shadaj, kevinb

Reviewed By: kevinb

Subscribers: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D28136