Allow cursor to be placed after fractions
Summary:
We need to add `\\frac` to the list of ignored commands, since it isn't supposed to be treated as a
command in the same sense as `\\log` or `\\cos`--that is, we need to let the cursor be placed after
it.

On Asana @ https://app.asana.com/0/132409062806753/135393867621649.

Test Plan:
- Open up the keypad.
- Type a fraction, and put some math after it (e.g., type `4/5 6`).
    - Verify that you can place your cursor between the fraction and the contents that follow.

Auditors: kevinb