Replace command blacklist with a command whitelist
Summary:
Previously, we were blacklisting certain commands so that we wouldn't treat them as functions
(i.e., so that we would allow the user to place the cursor after those nodes, since for fractions,
we prevent the cursor from being placed between the function name and its parentheses). However,
the blacklist kept expanding, as in 8284d60 and 477b077, and there seem to be even more commands
that we'll need to add, like `\\sqrt`. So, instead, I've replaced the blacklist with a whitelist of
known functions that we want to support, since that will be more robust and deliberate.

On Asana @ https://app.asana.com/0/132409062806753/143677267087842.

Test Plan:
- Verify that you can't place the cursor between 'cos', 'sin', 'tan', or 'log' and their parentheses.
- Verify that you _can_ place the cursor after any Greek symbols (Pi, Theta), radical sign, or set of parentheses.

Auditors: kevinb