Place cursor inside parens when dragging along a function
Summary:
Our existing logic for cursor placement in the case of dragging along a function (e.g., when
dragging the handle across the `cos` of `cos(45)`) was faulty in that it would continue rightward
until it hit the end of the input, rather than continuing rightward until the function name ended
(with a set of parens). This diff fixes it by explicitly searching for the set of parens that
follow the function name; in the process.

On Asana @ https://app.asana.com/0/132409062806753/143677267087851.

Test Plan:
- Open up the keypad.
- Add "123cos(45)6" to the input.
- Drag the cursor along the length of the input.
    - Verify that the cursor is placed at the beginning of the inside of the parens whenever you drag over "cos".
    - Verify that placement works on either side of the function, and within the function's parens, as expected.
- Run through the existing test suite (`npm run test`) to verify that existing transitions are unaffected.

Reviewers: shadaj, kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D28177