Don't assume that touches end over focused node
Summary:
In D27227, I introduced an optimization to avoid computing the node over which a touch ended, based
on the assumption that a touch would end over the focused node. This turned out to be an invalid
assumption given the existence of multi-touch interactions (I've outlined a sample problematic case
in the `TODO` below).

Long-term, we should migrate to properly tracking distinct gestures; but this will avoid buggy
behavior when, e.g., two keys are tapped in quick succession.

On Asana @ https://app.asana.com/0/132409062806753/144457874535376.

Test Plan:
- Open up the keypad on a phone.
- Touch down on a key (e.g., '4').
- Touch down on another key (e.g., '5'), without releasing.
- Release your first finger.
    - Verify that the keypad triggers '4' (and not '5').
- Release your second finger.
    - Verify that the keypad triggers '5'.

Reviewers: jared

Differential Revision: https://phabricator.khanacademy.org/D28346