Implement left- and right-parenthesis keys
Summary:
We now support inserting left and right parentheses. MathQuill doesn't _really_ seem to support
mismatched parens, so instead of treating these keys as "insert a single paren", we're piggybacking
on MathQuill's support for automatic parens insertion. Essentially, when you press the left- or
right-parenthesis key, we'll insert the left or right parenthesis, and then also insert a greyed
out parenthesis to match it, on the appropriate side of the nearest expression. Inserting any math
beyond a grayed out parenthesis will cause it to go dark, as well typing the single paren key at
its location.

Nefaur and I played around with the behavior, and we felt like this was a good place to start. We
may need to go back and tweak the behavior in the future.

On Asana @ https://app.asana.com/0/157213482411363/157213482411386.

Test Plan:
- Open up the keypad.
- Type a left parenthesis key.
    - Verify that both are inserted, and that your cursor ends up inside the parens.
    - Verify that deleting from inside the empty parens, even with a grayed-out right paren, deletes the expression.
    - Verify that deleting from outside the empty parens, even with a grayed-out right paren, selects the expression.
    - Verify that tapping beyond the right paren and typing more math causes it to go dark.
- Type some math. Type the right paren.
    - Verify that the parens appear around the expression to the left of the cursor.
    - Verify that navigating to the start of the expression and tapping the left paren key causes it to go dark.
    - Verify that navigating to the start of the expression and writing any math causes it to go dark.
    - Verify that navigating halfway through the expression and tapping the left paren key causes the ghost paren to move in and complete.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D29126