Reverse ordering of keypad pages
Summary:
On tablets, we want to show the 'first' page of the keypad on the right. In order to make the
layout consistent across devices, that means that we also want the 'first' page of the view pager
to be the rightmost page--that is, we now want the view pager to operate right-to-left, rather than
left-to-right.

On Asana @ https://app.asana.com/0/157213482411363/157213482411376.

Test Plan:
- Open up the keypad on a phone.
    - Verify that the number page is the initial page.
    - Verify that the rightmost icon in the pager indicator is active.
    - Verify that dragging left reveals the empty background.
    - Verify that dragging right reveals the advanced page.
    - Verify that you can drag to select the advanced page.
    - Verify that tapping any key on the advanced page transitions back to the basic page.
- Open up the keypad on a tablet.
    - Verify that the number page appears on the right.

{F548027}

{F548022}

{F548023}

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D29141