Render the keypad compactly when narrower than full-screen
Summary:
This diff implements the narrow or 'compact' keypad layout--that is, the layout for the case in
which the keypad does not take up the full width of the screen.

In general, this should be possible when:

- Any keypad is rendered in tablet landscape.
- The fraction keypad is rendered in tablet portrait.
- The fraction keypad is rendered in phone landscape.

One tricky bit was ensuring that the rounded corners properly cropped all child views. We can't use
`overflow: hidden` because it breaks the popover and echo animation behaviors, so, intead, I had to
propagate the border cropping all the way through the view heirarchy. Open to other ideas :(

On Asana @ https://app.asana.com/0/157213482411363/157333652496606.

Test Plan:
- Open up the fraction and expression keypads on a phone and tablet in portrait and landscape.
    - Verify that the keypad renders compactly in the cases enumerated in the summary.
    - Verify that, in each case, the corners are cropped properly.

{F548155}

{F548156}

{F548157}

{F548158}

{F548159}

{F548160}

Reviewers: jared, benkomalo, kevinb

Reviewed By: jared, benkomalo, kevinb

Differential Revision: https://phabricator.khanacademy.org/D29206