Avoid shrinking SVG icons on small viewports
Summary:
We want to render the focus state (the green box behind the icon) at a 4px inset. Previously, we
were achieving this effect by adding 4px of padding to the entire button, and setting the
background to the focus state color. However, this can unintentionally shrink down our icons, since
if the button is, say 50px-by-50px, and we apply 4px padding in each direction, then the icon will
be shrunk down to 42px-by-42px.

Instead, in order to achieve the desired layout constraints (icons are always at least
48px-by-48px, with the focus state always inset at 4px), we need to render the focus state as an
entirely separate view, rendered below the icon.

On Asana @ https://app.asana.com/0/157213482411363/159462668773540.

Test Plan:
- Open up the keypad with mobile emulation.
- Make the screen narrower than 280px.
    - Verify that the buttons are rendered to fill the screen.
    - Verify that the underlying SVG icon is 48px-by-48px, centered within the larger icon.
    - Verify that the focus state is inset at 4px from the edges, and that it can be smaller than the SVG icon.
    - Verify that the echo animation behavior is as-before.

As an example: {F548675}

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D29306