Accept style overrides for keypad container
Summary:
This will allow webapp (via Perseus) to override some of the styles on the keypad. In particular,
it'll be able to move the keypad up the page to sit atop the 'Check answer' bar.

Test Plan: Add some arbitrary styles via the playground; verify that they're applied properly.

Auditors: kevinb